Add virtualPoints creating utils
