+++ src/components/TextElementResizer.tsx
@@ -3,11 +3,7 @@ import {
   VisualizerMachineContext,
   VisualizerTextElement,
 } from "../machines/visualizerMachine";
-import {
-  calculateElementAbsolutePoint,
-  calculateElementViewportSize,
-  convertToViewportPoint,
-} from "../utils";
+import { createDiagonalDirectionVirtualPoints } from "../utils";
 import { DiagonalDirection } from "./GenericElementResizer";
 import VirtualPoint from "./VirtualPoint";
 
@@ -27,49 +23,12 @@ function TextElementResizer({
     direction: DiagonalDirection
   ) => void;
 }) {
-  const absolutePoint = calculateElementAbsolutePoint(textElement);
-  const elementViewportPoint = convertToViewportPoint({
-    canvasPoint: {
-      x: absolutePoint.minX,
-      y: absolutePoint.minY,
-    },
+  const virtualPoints = createDiagonalDirectionVirtualPoints({
+    element: textElement,
     devicePixelRatio,
     origin,
     zoom,
   });
-
-  const elementViewportSize = calculateElementViewportSize({
-    element: textElement,
-    devicePixelRatio,
-    zoom,
-  });
-
-  const virtualPoints: {
-    direction: DiagonalDirection;
-    left: number;
-    top: number;
-  }[] = [
-    {
-      direction: "up-left",
-      left: elementViewportPoint.x,
-      top: elementViewportPoint.y,
-    },
-    {
-      direction: "up-right",
-      left: elementViewportPoint.x + elementViewportSize.width,
-      top: elementViewportPoint.y,
-    },
-    {
-      direction: "down-left",
-      left: elementViewportPoint.x,
-      top: elementViewportPoint.y + elementViewportSize.height,
-    },
-    {
-      direction: "down-right",
-      left: elementViewportPoint.x + elementViewportSize.width,
-      top: elementViewportPoint.y + elementViewportSize.height,
-    },
-  ];
 
   return (
     <>
